pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
src_paths = ["cafeteria", "test"]
known_first_party = "cafeteria"

[tool.poetry]
name = "cafeteria"
version = "0.22.2a0"
repository = "https://github.com/abn/cafeteria"
description = "Cafeteria: A convenience package providing various building blocks enabling pythonic patterns."
authors = [
    "Arun Babu Neelicattu <arun.neelicattu@gmail.com>",
    "Josha Inglis <joshainglis@gmail.com>",
    "Betsy Alpert <lizbeth.alpert@gmail.com>"
]
license = "APLv2"
classifiers = [
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
]
keywords = []
readme = "README.rst"


[tool.poetry.dependencies]
python = "^3.8"
PyYAML = ">=3.13,!=6.0.0,!=5.4.0,!=5.4.1"

[tool.poetry.extras]
# PyYAML is only required when YAML logging configuration is used.
yaml = ["PyYAML"]

[tool.poetry.dev-dependencies]
pre-commit = {version = "^2.7", python = ">=3.6.1"}
flake8 = "^3.8"
coverage = "^5.2"
black = { version = "^21.5b0", allow-prereleases = true, python = ">=3.6.2,<4.0" }
tox = "^3.20"
pytest = "^6.0"
pytest-cov = "^2.7"
